requirements.txt
# Core dependencies - required for basic functionality
requests>=2.31.0
colorama>=0.4.6
python-dateutil>=2.8.2
tabulate>=0.9.0

# Optional dependencies for advanced features
python-evtx>=0.8.0  # For Windows EVTX parsing
pandas>=1.5.0       # For data analysis
xmltodict>=0.13.0   # For XML parsing
elasticsearch>=8.0.0  # For Elasticsearch integration

# GUI dependencies (optional)
# tkinter is built-in with Python on Windows

# Development dependencies (optional - for code quality)
pytest>=7.0.0       # For testing
flake8>=6.0.0       # For code linting
# black - install manually if needed: pip install black
